Added unit tests, updated docs

+++ src/boundary/FT_bc.f90
@@ -3,10 +3,6 @@ module FT_bc
   !< Apply flow tangency boundary condition
   !--------------------------------------------
   use vartypes
-!  use global_vars, only: xn
-!  use global_vars, only: yn
-!  use global_vars, only: zn
-
   use copy_bc   , only : copy3
 
   implicit none
@@ -20,7 +16,7 @@ module FT_bc
       !< Apply flow tangency boundary condition
       implicit none
       type(extent), intent(in) :: dims
-      real, dimension(-2:dims%imx+2, -2:dims%jmx+2, -2:dims%kmx+2, 1:dims%n_var), intent(inout):: qp
+      real(wp), dimension(-2:dims%imx+2, -2:dims%jmx+2, -2:dims%kmx+2, 1:dims%n_var), intent(inout):: qp
       character(len=*), intent(in) :: face
       !< Face over which flow tangency condition has to be applied
       type(facetype), dimension(-2:dims%imx+3,-2:dims%jmx+2,-2:dims%kmx+2), intent(in) :: Ifaces
@@ -29,10 +25,7 @@ module FT_bc
       !< Input varaible which stores J faces' area and unit normal
       type(facetype), dimension(-2:dims%imx+2,-2:dims%jmx+2,-2:dims%kmx+3), intent(in) :: Kfaces
       !< Input varaible which stores K faces' area and unit normal
-!      real :: dot1
-!      real :: dot2
-!      real :: dot3
-      real :: dot
+      real(wp) :: dot
       integer :: i
       integer :: j
       integer :: k
@@ -52,12 +45,6 @@ module FT_bc
                 qp(-l+1,j,k,2) = qp(l,j,k,2) - (2.0*dot*Ifaces(1,j,k)%nx)
                 qp(-l+1,j,k,3) = qp(l,j,k,3) - (2.0*dot*Ifaces(1,j,k)%ny)
                 qp(-l+1,j,k,4) = qp(l,j,k,4) - (2.0*dot*Ifaces(1,j,k)%nz)
-!              dot1=sum(qp(1,j,k,2:4)*xn(1,j,k,1:3))
-!              dot2=sum(qp(2,j,k,2:4)*xn(1,j,k,1:3))
-!              dot3=sum(qp(3,j,k,2:4)*xn(1,j,k,1:3))
-!              qp( 0,j,k,2:4) = qp(1,j,k,2:4) - (2*dot1*xn(1,j,k,1:3))
-!              qp(-1,j,k,2:4) = qp(2,j,k,2:4) - (2*dot2*xn(1,j,k,1:3))
-!              qp(-2,j,k,2:4) = qp(3,j,k,2:4) - (2*dot3*xn(1,j,k,1:3))
               end do
             end do
           end do
@@ -69,12 +56,6 @@ module FT_bc
                 qp(imx+l-1,j,k,2) = qp(imx-l,j,k,2) - (2.0*dot*Ifaces(imx,j,k)%nx)
                 qp(imx+l-1,j,k,3) = qp(imx-l,j,k,3) - (2.0*dot*Ifaces(imx,j,k)%ny)
                 qp(imx+l-1,j,k,4) = qp(imx-l,j,k,4) - (2.0*dot*Ifaces(imx,j,k)%nz)
-!              dot1=sum(qp(imx-1,j,k,2:4)*xn(imx,j,k,1:3))
-!              dot2=sum(qp(imx-2,j,k,2:4)*xn(imx,j,k,1:3))
-!              dot3=sum(qp(imx-3,j,k,2:4)*xn(imx,j,k,1:3))
-!              qp(imx+0,j,k,2:4) = qp(imx-1,j,k,2:4) - (2*dot1*xn(imx,j,k,1:3))
-!              qp(imx+1,j,k,2:4) = qp(imx-2,j,k,2:4) - (2*dot2*xn(imx,j,k,1:3))
-!              qp(imx+2,j,k,2:4) = qp(imx-3,j,k,2:4) - (2*dot3*xn(imx,j,k,1:3))
               end do
             end do
           end do
@@ -86,12 +67,6 @@ module FT_bc
                 qp(i,-l+1,k,2) = qp(i,l,k,2) - (2.0*dot*Ifaces(i,1,k)%nx)
                 qp(i,-l+1,k,3) = qp(i,l,k,3) - (2.0*dot*Ifaces(i,1,k)%ny)
                 qp(i,-l+1,k,4) = qp(i,l,k,4) - (2.0*dot*Ifaces(i,1,k)%nz)
-!              dot1=sum(qp(i,1,k,2:4)*yn(i,1,k,1:3))
-!              dot2=sum(qp(i,2,k,2:4)*yn(i,1,k,1:3))
-!              dot3=sum(qp(i,3,k,2:4)*yn(i,1,k,1:3))
-!              qp(i, 0,k,2:4) = qp(i,1,k,2:4) - (2*dot1*yn(i,1,k,1:3))
-!              qp(i,-1,k,2:4) = qp(i,2,k,2:4) - (2*dot2*yn(i,1,k,1:3))
-!              qp(i,-2,k,2:4) = qp(i,3,k,2:4) - (2*dot3*yn(i,1,k,1:3))
               end do
             end do
           end do
@@ -103,12 +78,6 @@ module FT_bc
                 qp(i,jmx+l-1,k,2) = qp(i,jmx-l,k,2) - (2.0*dot*Ifaces(i,jmx,k)%nx)
                 qp(i,jmx+l-1,k,3) = qp(i,jmx-l,k,3) - (2.0*dot*Ifaces(i,jmx,k)%ny)
                 qp(i,jmx+l-1,k,4) = qp(i,jmx-l,k,4) - (2.0*dot*Ifaces(i,jmx,k)%nz)
-!              dot1=sum(qp(i,jmx-1,k,2:4)*yn(i,jmx,k,1:3))
-!              dot2=sum(qp(i,jmx-2,k,2:4)*yn(i,jmx,k,1:3))
-!              dot3=sum(qp(i,jmx-3,k,2:4)*yn(i,jmx,k,1:3))
-!              qp(i,jmx+0,k,2:4) = qp(i,jmx-1,k,2:4) - (2*dot1*yn(i,jmx,k,1:3))
-!              qp(i,jmx+1,k,2:4) = qp(i,jmx-2,k,2:4) - (2*dot2*yn(i,jmx,k,1:3))
-!              qp(i,jmx+2,k,2:4) = qp(i,jmx-3,k,2:4) - (2*dot3*yn(i,jmx,k,1:3))
               end do
             end do
           end do
@@ -120,12 +89,6 @@ module FT_bc
                 qp(i,j,-l+1,2) = qp(i,j,l,2) - (2.0*dot*Ifaces(i,j,1)%nx)
                 qp(i,j,-l+1,3) = qp(i,j,l,3) - (2.0*dot*Ifaces(i,j,1)%ny)
                 qp(i,j,-l+1,4) = qp(i,j,l,4) - (2.0*dot*Ifaces(i,j,1)%nz)
-!              dot1=sum(qp(i,j,1,2:4)*zn(i,j,1,1:3))
-!              dot2=sum(qp(i,j,2,2:4)*zn(i,j,1,1:3))
-!              dot3=sum(qp(i,j,3,2:4)*zn(i,j,1,1:3))
-!              qp(i,j, 0,2:4) = qp(i,j,1,2:4) - (2*dot1*zn(i,j,1,1:3))
-!              qp(i,j,-1,2:4) = qp(i,j,2,2:4) - (2*dot2*zn(i,j,1,1:3))
-!              qp(i,j,-2,2:4) = qp(i,j,3,2:4) - (2*dot3*zn(i,j,1,1:3))
               end do
             end do
           end do
@@ -137,12 +100,6 @@ module FT_bc
                 qp(i,j,kmx+l-1,2) = qp(i,j,kmx-l,2) - (2.0*dot*Ifaces(i,j,kmx)%nx)
                 qp(i,j,kmx+l-1,3) = qp(i,j,kmx-l,3) - (2.0*dot*Ifaces(i,j,kmx)%ny)
                 qp(i,j,kmx+l-1,4) = qp(i,j,kmx-l,4) - (2.0*dot*Ifaces(i,j,kmx)%nz)
-!              dot1=sum(qp(i,j,kmx-1,2:4)*zn(i,j,kmx,1:3))
-!              dot2=sum(qp(i,j,kmx-2,2:4)*zn(i,j,kmx,1:3))
-!              dot3=sum(qp(i,j,kmx-3,2:4)*zn(i,j,kmx,1:3))
-!              qp(i,j,kmx+0,2:4) = qp(i,j,kmx-1,2:4) - (2*dot1*zn(i,j,kmx,1:3))
-!              qp(i,j,kmx+1,2:4) = qp(i,j,kmx-2,2:4) - (2*dot2*zn(i,j,kmx,1:3))
-!              qp(i,j,kmx+2,2:4) = qp(i,j,kmx-3,2:4) - (2*dot3*zn(i,j,kmx,1:3))
               end do
             end do
           end do
