+++ src/read_write/write/write_output_tec_node.f90
@@ -40,27 +40,27 @@ module write_output_tec_node
   integer :: i,j,k
   character(len=*), parameter :: format="(35e25.15)"
   integer :: imx, jmx, kmx
-  real, dimension(:, :, :), pointer :: density      
+  real(wp), dimension(:, :, :), pointer :: density      
    !< Rho pointer, point to slice of qp (:,:,:,1)
-  real, dimension(:, :, :), pointer :: x_speed      
+  real(wp), dimension(:, :, :), pointer :: x_speed      
    !< U pointer, point to slice of qp (:,:,:,2) 
-  real, dimension(:, :, :), pointer :: y_speed      
+  real(wp), dimension(:, :, :), pointer :: y_speed      
    !< V pointer, point to slice of qp (:,:,:,3) 
-  real, dimension(:, :, :), pointer :: z_speed      
+  real(wp), dimension(:, :, :), pointer :: z_speed      
    !< W pointer, point to slice of qp (:,:,:,4)
-  real, dimension(:, :, :), pointer :: pressure     
+  real(wp), dimension(:, :, :), pointer :: pressure     
    !< P pointer, point to slice of qp (:,:,:,5)
-  real, dimension(:, :, :), pointer :: tk        
+  real(wp), dimension(:, :, :), pointer :: tk        
   !< TKE, point to slice of qp (:,:,:,6)
-  real, dimension(:, :, :), pointer :: tw        
+  real(wp), dimension(:, :, :), pointer :: tw        
   !< Omega, point to slice of qp (:,:,:,7)
-  real, dimension(:, :, :), pointer :: te        
+  real(wp), dimension(:, :, :), pointer :: te        
   !< Dissipation, point to slice of qp (:,:,:,7)
-  real, dimension(:, :, :), pointer :: tv        
+  real(wp), dimension(:, :, :), pointer :: tv        
   !< SA visocity, point to slice of qp (:,:,:,6)
-  real, dimension(:, :, :), pointer :: tkl       
+  real(wp), dimension(:, :, :), pointer :: tkl       
   !< KL K-KL method, point to slice of qp (:,:,:,7)
-  real, dimension(:, :, :), pointer :: tgm       
+  real(wp), dimension(:, :, :), pointer :: tgm       
   !< Intermittency of LCTM2015, point to slice of qp (:,:,:,8)
   public :: write_file
 
@@ -74,7 +74,7 @@ module write_output_tec_node
       type(schemetype), intent(in) :: scheme
       type(extent), intent(in) :: dims
       type(nodetype), dimension(-2:dims%imx+3,-2:dims%jmx+3,-2:dims%kmx+3), intent(in) :: nodes 
-      real, dimension(-2:dims%imx+2, -2:dims%jmx+2, -2:dims%kmx+2, 1:dims%n_var), intent(in), target :: state
+      real(wp), dimension(-2:dims%imx+2, -2:dims%jmx+2, -2:dims%kmx+2, 1:dims%n_var), intent(in), target :: state
       integer :: n
       character(len=*), parameter :: err="Write error: Asked to write non-existing variable- "
 
@@ -301,7 +301,7 @@ module write_output_tec_node
       !< Write scalar variable in the output file
       implicit none
       integer, intent(in) :: index
-      real, dimension(index:imx-index,index:jmx-index,index:kmx-index), intent(in) :: var
+      real(wp), dimension(index:imx-index,index:jmx-index,index:kmx-index), intent(in) :: var
       character(len=*),       intent(in):: name
 
       DebugCall('write_scalar'//trim(name))
