+++ tests/test_weno_NM.f90
@@ -6,10 +6,10 @@ integer function test_weno_NM() result(r)
    type(celltype), dimension(-2:7,-2:4,-2:4)  :: cells
    !imx,jmx,kmx, = 5,2,2
    !n_var = 5
-   real, dimension(-2:7,-2:4,-2:4,1:5) :: qp=0.0
-   real, dimension(0:6,1:1,1:1,5) :: x_qp_left, x_qp_right
-   real, dimension(1:4,0:3,1:1,5) :: y_qp_left, y_qp_right
-   real, dimension(1:4,1:1,0:3,5) :: z_qp_left, z_qp_right
+   real(wp), dimension(-2:7,-2:4,-2:4,1:5) :: qp=0.0
+   real(wp), dimension(0:6,1:1,1:1,5) :: x_qp_left, x_qp_right
+   real(wp), dimension(1:4,0:3,1:1,5) :: y_qp_left, y_qp_right
+   real(wp), dimension(1:4,1:1,0:3,5) :: z_qp_left, z_qp_right
 
    r = 1
